Refactor PaginationComponent component file structure
