+++ src/components/TabLists/PaginationComponent/PaginationComponent.js
@@ -1,56 +1,12 @@
 /* eslint-disable no-underscore-dangle */
 import React from 'react';
 import PropTypes from 'prop-types';
-import { injectIntl, intlShape } from 'react-intl';
 import {
-  withStyles, Button, Typography,
+  Button,
 } from '@material-ui/core';
 import { ArrowBackIos, ArrowForwardIos } from '@material-ui/icons';
 import Container from '../../Container';
-import { keyboardHandler } from '../../../utils';
-import styles from './styles';
-
-// Page number element
-const PageElement = ({
-  intl, isActive, number, onClick, ...rest
-}) => (
-  <li>
-    <Typography
-      variant="subtitle1"
-      component="p"
-      style={{
-        color: isActive ? 'red' : 'black',
-        cursor: isActive ? 'auto' : 'pointer',
-        textDecoration: isActive ? 'underline' : 'none',
-      }}
-      role={!isActive ? 'link' : null}
-      onClick={!isActive ? onClick : null}
-      onKeyPress={!isActive ? keyboardHandler(onClick, ['space', 'enter']) : null}
-      tabIndex={isActive ? null : '0'}
-      {...rest}
-    >
-      <Typography variant="srOnly">
-        {
-          isActive
-            ? intl.formatMessage({ id: 'general.pagination.currentlyOpenedPage' }, { count: number })
-            : intl.formatMessage({ id: 'general.pagination.openPage' }, { count: number })
-        }
-      </Typography>
-      <span aria-hidden="true">
-        {number}
-      </span>
-    </Typography>
-  </li>
-
-);
-
-PageElement.propTypes = {
-  intl: intlShape.isRequired,
-  isActive: PropTypes.bool.isRequired,
-  number: PropTypes.number.isRequired,
-  onClick: PropTypes.func.isRequired,
-};
-
+import PageElement from './PageElement';
 
 class PaginationComponent extends React.Component {
   constructor(props) {
@@ -156,4 +112,4 @@ PaginationComponent.defaultProps = {
   maxShownPages: 7,
 };
 
-export default withStyles(styles)(injectIntl(PaginationComponent));
+export default PaginationComponent;
