+++ src/components/TabLists/PaginationComponent/index.js
@@ -1,3 +1,6 @@
+import { withStyles } from '@material-ui/core';
+import { injectIntl } from 'react-intl';
 import PaginationComponent from './PaginationComponent';
+import styles from './styles';
 
-export default PaginationComponent;
+export default withStyles(styles)(injectIntl(PaginationComponent));
